requirements.txt
requests==2.31.0
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
validators==0.22.0
colorama==0.4.6
tabulate==0.9.0
click==8.1.7
beautifulsoup4==4.12.2